+++ libft.h
@@ -6,7 +6,7 @@
 /*   By: vgallois <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/07 15:05:50 by vgallois          #+#    #+#             */
-/*   Updated: 2019/11/07 09:36:40 by vgallois         ###   ########.fr       */
+/*   Updated: 2019/12/10 17:50:42 by vgallois         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -60,10 +60,10 @@ void			ft_putchar_fd(char c, int fd);
 void			ft_putnstr(char *s, int n);
 size_t			ft_putstr(char *s);
 void			ft_putstr_fd(char *s, int fd);
-void			ft_putnbr_fd(int n, int fd);
+void			ft_putnbr_fd(intmax_t n, int fd);
 void			ft_putendl_fd(char *s, int fd);
 void			ft_putendl(char *s);
-void			ft_putnbrbase(int nb, char *b);
+void			ft_putnbrbase(intmax_t nb, char *b);
 t_list			*ft_lstnew(void const *content);
 void			ft_lstadd_front(t_list **alst, t_list *new);
 int				ft_lstsize(t_list *lst);
